Extract pull request integration tests
