+++ docs/DECISIONS.md
@@ -266,3 +266,4 @@
 - Message feedback and artifact preview coverage should stay out of `WorkspaceModelTests`. Feedback persistence lives in `WorkspaceFeedbackIntegrationTests` because it crosses model mutation, thread events, and transcript surfaces. Pure `ToolArtifactState` image/document/text preview derivation lives in `QuillCodeToolCardSurfaceTests` because it is value-type surface behavior independent of workspace state.
 - Runtime issue integration coverage lives in `WorkspaceRuntimeIssueIntegrationTests`, not `WorkspaceModelTests`. Applying runtime status, projecting runtime issues into top-bar/settings surfaces, diagnostic redaction, and retry-last-turn mutation cross runtime state, workspace surfaces, commands, and composer state; keep pure issue construction in `WorkspaceRuntimeIssueBuilderTests` and recovery-action mapping in `QuillCodeRuntimeIssueRecoveryPlannerTests`.
 - Thread lifecycle integration coverage lives in `WorkspaceThreadLifecycleIntegrationTests`, not `WorkspaceModelTests`. New chat selection, fork/compact context, project-thread fallback selection, pin/archive persistence, and rename/duplicate/unarchive/delete flows cross the workspace model, stores, sidebar ordering, selected project, and top-bar state; keep pure mutation algorithms in `WorkspaceThreadLifecycleEngineTests`.
+- Pull request command integration coverage lives in `WorkspacePullRequestIntegrationTests`, not `WorkspaceModelTests`. Remote PR view/checks/diff execution through SSH, slash `/pr` command dispatch, and PR command composer prefills cross workspace commands, slash parsing, SSH remote execution, tool cards, and composer state; keep pure GitHub CLI argument construction in tool tests and remote git builder tests.
